full code cleanup
